assistant2: Extract method for adding a new profile to the settings
